cleanup some things and moved some functions to utils
